feat: Add getUsers route and controller for retrieving all users
This commit adds a new route and controller in the API for retrieving all users. The getUsers route is added to the user.route.js file, and the getUsers controller function is implemented in the user.controller.js file. This functionality allows users with admin privileges to view all users, including their details such as username, email, and registration date. The response includes the list of users, the total number of users, and the number of users created in the last month.